apps/api/requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-decouple==3.8
redis==7.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
